backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
google-generativeai==0.3.2
httpx==0.25.2
python-multipart==0.0.6
cors==1.0.1
fastapi-cors==0.0.6

